feat: Fetch and populate dropdowns with student and program data
This commit adds functionality to fetch and populate the dropdown menus for selecting a student and a program. The code fetches the data from the respective API endpoints and dynamically creates the options for each dropdown. This feature enhances the user experience by providing an up-to-date list of students and programs to choose from when updating a student's information.